requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
black==24.8.0
ca-certificates==2024.11.26
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.1.1
cycler==0.12.1
detectron2==0.6
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.55.2
fsspec==2024.2.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
gpytorch==1.13
grpcio==1.68.1
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
iopath==0.1.9
jaxtyping==0.2.19
jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
lazy-loader==0.4
libffi==3.4.4
libgcc-ng==11.2.0
libgomp==11.2.0
libstdcxx-ng==11.2.0
linear-operator==0.5.3
loguru==0.7.3
markdown==3.7
markupsafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
mypy-extensions==1.0.0
ncurses==6.4
netcal==1.3.6
networkx==3.0
ninja==1.11.1.2
numpy==1.23.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.8.1
onnx-simplifier==0.3.5
onnxoptimizer==0.3.13
onnxruntime==1.8.0
opencv-python==4.10.0.84
openssl==3.0.15
opt-einsum==3.4.0
packaging==24.2
pathspec==0.12.1
pillow==10.2.0
pip==24.2
platformdirs==4.3.6
portalocker==3.0.0
protobuf==3.20.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycocotools==2.0.7
pyparsing==3.1.4
pyro-api==0.1.2
pyro-ppl==1.9.1
python==3.8.20
pywavelets==1.4.1
pyyaml==6.0.2
readline==8.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
setuptools==75.1.0
six==1.17.0
sqlite==3.45.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tikzplotlib==0.9.8
tk==8.6.14
tomli==2.2.1
tqdm==4.67.1
triton==2.2.0
typeguard==4.4.0
typing-extensions==4.12.2
urllib3==2.2.3
werkzeug==3.0.6
wheel==0.44.0
xz==5.4.6
yacs==0.1.8
yolox==0.3.0
zipp==3.20.2
zlib==1.2.13
